backend/audio/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27
pydantic>=2.6
python-multipart>=0.0.9
google-cloud-aiplatform>=1.63.0
google-cloud-storage>=2.14.0
google-auth>=2.27
protobuf>=4.25.3
grpcio>=1.64
packaging>=23.2